Using Controllers for post.js and user.js
