+++ backend/routes/posts.js
@@ -1,7 +1,8 @@
 const express = require("express");
 const multer = require("multer");
 
-const Post = require("../models/post");
+const PostController = require("../controllers/post");
+
 const checkAuth = require("../middleware/check-auth");
 
 const router = express.Router();
@@ -31,115 +32,14 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post(
-  	"",
-	checkAuth,
-  	multer({ storage: storage }).single("image"),
-  		(req, res, next) => {
-    			const url = req.protocol + "://" + req.get("host");
-    			const post = new Post({
-      				title: req.body.title,
-      				content: req.body.content,
-      				imagePath: url + "/images/" + req.file.filename,
-				creator: req.userData.userId
-    			});
-    			post.save().then(createdPost => {
-      			res.status(201).json({
-        			message: "Post added successfully",
-        			post: {
-          				...createdPost,
-          				id: createdPost._id
-        			}
-      			});
-    		}).catch(error => {
-			res.status(500).json({
-				message: "Creating a post Failed!"
-			});
-		});
-  	}
-);
+router.post("", checkAuth, multer({ storage: storage }).single("image"), PostController.createPost);
 
-router.put(
-  "/:id",
-	checkAuth,
-  multer({ storage: storage }).single("image"),
-  (req, res, next) => {
-    let imagePath = req.body.imagePath;
-    if (req.file) {
-      const url = req.protocol + "://" + req.get("host");
-      imagePath = url + "/images/" + req.file.filename
-    }
-    const post = new Post({
-      _id: req.body.id,
-      title: req.body.title,
-      content: req.body.content,
-      imagePath: imagePath,
-	creator: req.userData.userId
-    });
-    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
-	if(result.nModified > 0){
-		res.status(200).json({ message: "Update successful!" });
-	} else {
-		res.status(401).json({ message: "Not Authorized"});
-	}
-    }).catch(error => {
-		res.status(500).json({
-			message: "Couldn't update post!"
-		});
-	});
-  }
-);
+router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), PostController.updatePost);
 
-router.get("", (req, res, next) => {
-  const pageSize = +req.query.pagesize;
-  const currentPage = +req.query.page;
-  const postQuery = Post.find();
-  let fetchedPosts;
-  if(pageSize && currentPage){
-	postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
-  }
-  postQuery.find().then(documents => {
-	fetchedPosts = documents;
-	return Post.countDocuments();
-  }).then(count => {
-	res.status(200).json({
-	message: "Posts fetched successfully!",
-	posts: fetchedPosts,
-	maxPosts: count
-	});
-  }).catch(error => {
-		res.status(500).json({
-			message: "Fetching posts failed!"
-		});
-	});
-});
+router.get("", PostController.fetchPosts);
 
-router.get("/:id", (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
-    if (post) {
-      res.status(200).json(post);
-    } else {
-      res.status(404).json({ message: "Post not found!" });
-    }
-  }).catch(error => {
-                res.status(500).json({
-                        message: "Fetching posts failed!"
-                });
-        });
-});
+router.get("/:id", PostController.findPostsById);
 
-router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
-        if(result.n > 0){
-                res.status(200).json({ message: "Deletion successful!" });
-        } else {
-                res.status(401).json({ message: "Not Authorized"});
-        }
-  }).catch(error => {
-                res.status(500).json({
-                        message: "Fetching posts failed!"
-                });
-        });
-});
+router.delete("/:id", checkAuth, PostController.deletePosts);
 
 module.exports = router;
